refactor week4
