+++ src/hotel_reservations/synthetic_data.py
